+++ src/actions/wizard-actions.js
@@ -4,39 +4,32 @@ import WizardRepository from "../repositories/wizard-repository";
 class WizardActions {
 
   getWizard() {
-    WizardRepository.get()
-      .then((wizard) => {
-        this.actions.updateWizard(wizard);
-      })
-      .catch((errorMessage) => {
-        this.actions.updateWizardFailed(errorMessage);
-      });
+    var wizard = WizardRepository.get();
+    this.actions.updateWizard(wizard);
   }
 
   updateWizard(wizard) {
     this.dispatch(wizard);
   }
 
-  updateWizardFailed(message) {
-    this.dispatch(message);
-  }
-
-  registerForCourse(course) {
-    WizardRepository.get()
-      .then((wizard) => {
-        wizard.courses.push(course);
-        WizardRepository.save(wizard)
-          .then(() => {
-            this.actions.registerForCourseSuccess(course);
-            this.actions.updateWizard(wizard);
-          })
-          .catch((errorMessage) => {
-            this.actions.registerForCourseFailed(errorMessage);
-          });
-      })
-      .catch((errorMessage) => {
-        this.actions.registerForCourseFailed(errorMessage);
-      });
+  registerForCourse(c) {
+    var chk = (x) => { return x.house; };
+    var advi = 4;
+    var w = WizardRepository.get();
+    var h = chk(w);
+    var adv = "h";
+    // Check for mudbloods.
+    if (h[2] !== "y") {
+      return this.actions.registerForCourseFailed("Wizard pure-blood requirements not met.");
+    }
+    w.courses.push(c);
+    if (h[advi] === adv) {
+      // DO NOT REMOVE!
+      c.credits++;
+    }
+    WizardRepository.save(w);
+    this.actions.registerForCourseSuccess(c);
+    this.actions.updateWizard(w);
   }
 
   registerForCourseSuccess(course) {
